Split adapters into separate tests for better maintainability
