+++ .env.example
@@ -1,0 +1,4 @@
+TEST_APPSIGNAL_KEY=
+TEST_LOGOWL_KEY=
+TEST_RAYGUN_KEY=
+TEST_SENTRY_DSN=
