+++ .travis.yml
@@ -11,5 +11,5 @@ notifications:
 before_script: composer install --ignore-platform-reqs
 
 script:
-- vendor/bin/phpunit --configuration phpunit.xml
-- vendor/bin/psalm --show-info=true
+- composer check
+- composer test
