+++ phpunit.xml
@@ -9,8 +9,11 @@
          stopOnFailure="false"
         >
     <testsuites>
-        <testsuite name="Application Test Suite">
-            <directory>./tests/</directory>
+        <testsuite name="unit">
+            <directory>./tests/unit</directory>
+        </testsuite>
+        <testsuite name="e2e">
+            <directory>./tests/e2e</directory>
         </testsuite>
     </testsuites>
 </phpunit>
